requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
blinker==1.6.3
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.2.1
cryptography==45.0.4
cssselect==1.3.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h11==0.14.0
h2==4.2.0
h5py==3.11.0
hpack==4.1.0
httpcore==1.0.5
httpx==0.27.2
humanfriendly==10.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
kaitaistruct==0.10
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
onnx==1.16.2
onnxruntime==1.19.0
openai==1.43.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parsel==1.10.0
pillow==10.4.0
Protego==0.4.0
protobuf==4.25.4
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
pydivert==2.1.0
Pygments==2.18.0
pyOpenSSL==25.1.0
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.2
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
rich==13.8.0
scipy==1.14.1
Scrapy==2.13.2
seaborn==0.13.2
selenium==4.33.0
selenium-wire==5.1.0
service-identity==24.2.0
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
tldextract==5.3.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
typing_extensions==4.13.2
tzdata==2024.1
ultralytics==8.2.80
ultralytics-thop==2.0.5
undetected-chromedriver==3.5.5
urllib3==2.4.0
w3lib==2.3.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
zope.interface==7.2
zstandard==0.23.0
aiohttp==3.9.5
scrapy-playwright==0.0.34
playwright==1.40.0